Editing LHTTP to work with older firmware
Turns out that the current LHTTP library has been designed to work with firmware version 2.5 and above. I'm going to implement a bunch of changes to strip the library of extrenous functionality to see how it goes... still at the proof of concept stage.

Removed QDN (domain resolution) from the code.